% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_bigpls_cox.R
\name{align_big_plscox}
\alias{align_big_plscox}
\title{Align a GD fit to a PLS fit (optional refit)}
\usage{
align_big_plscox(fit_gd, fit_pls, time, status, rotate = TRUE, refit = TRUE)
}
\arguments{
\item{fit_gd}{object from big_pls_cox_gd()}

\item{fit_pls}{object from big_pls_cox()}

\item{time, status}{Surv parts used for refit}

\item{rotate}{logical; Procrustes-rotate GD scores to the PLS basis}

\item{refit}{logical; refit a Cox model on the rotated GD scores}
}
\value{
fit_gd with $scores/$coefficients/$cox_fit possibly updated
}
\description{
Align a GD fit to a PLS fit (optional refit)
}
