% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_fast.R
\name{predict.big_pls_cox_fast}
\alias{predict.big_pls_cox_fast}
\title{Predictions for fast big PLS–Cox fits}
\usage{
\method{predict}{big_pls_cox_fast}(
  object,
  newdata = NULL,
  type = c("link", "risk", "response", "components"),
  comps = NULL,
  coef = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{"big_pls_cox_fast"}.}

\item{newdata}{Optional matrix or \code{big.matrix} with predictors.}

\item{type}{Type of prediction: \code{"link"}, \code{"risk"},
\code{"response"}, or \code{"components"}.}

\item{comps}{Optional components to use. Defaults to all.}

\item{coef}{Optional coefficient vector. Defaults to \code{object$coefficients}.}

\item{...}{Ignored.}
}
\description{
Predictions for fast big PLS–Cox fits
}
