% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-big_pls_cox.R
\name{print.big_pls_cox}
\alias{print.big_pls_cox}
\title{Print method for big_pls_cox objects}
\usage{
\method{print}{big_pls_cox}(x, ...)
}
\arguments{
\item{x}{An object of class \code{big_pls_cox}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for big_pls_cox objects
}
