% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_fast.R
\name{print.summary.big_pls_cox_fast}
\alias{print.summary.big_pls_cox_fast}
\title{Print method for summary.big_pls_cox_fast objects}
\usage{
\method{print}{summary.big_pls_cox_fast}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.big_pls_cox_fast}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for summary.big_pls_cox_fast objects
}
