% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-big_pls_cox.R
\name{summary.big_pls_cox}
\alias{summary.big_pls_cox}
\title{Summary for big_pls_cox objects}
\usage{
\method{summary}{big_pls_cox}(object, digits = 3, ...)
}
\arguments{
\item{object}{A \code{big_pls_cox} object.}

\item{digits}{Number of digits to print.}

\item{...}{Unused, for S3 compatibility.}
}
\description{
Summary for big_pls_cox objects
}
