% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_fast.R
\name{summary.big_pls_cox_fast}
\alias{summary.big_pls_cox_fast}
\title{Summary for big_pls_cox objects}
\usage{
\method{summary}{big_pls_cox_fast}(object, ...)
}
\arguments{
\item{object}{A \code{big_pls_cox} object.}

\item{...}{Not used.}
}
\description{
Summary for big_pls_cox objects
}
